+++ src/Services/Catalog/Application/IntegrationEvents/ProductFulfillmentMethodChangedToFbbIntegrationEvent.cs
@@ -1,4 +1,4 @@
 ﻿namespace Bazaar.Catalog.Application.IntegrationEvents;
 
 public record ProductFulfillmentMethodChangedToFbbIntegrationEvent(
-    string ProductId, decimal ProductLengthInCm, decimal ProductWidthInCm, decimal ProductHeightInCm, string SellerId) : IntegrationEvent;
+    string ProductId, float ProductLengthInCm, float ProductWidthInCm, float ProductHeightInCm, string SellerId) : IntegrationEvent;
