+++ src/Services/Catalog/Domain/Entities/CatalogItem.cs
@@ -8,9 +8,9 @@ public class CatalogItem
     public string ProductDescription { get; private set; }
     public decimal Price { get; private set; }
     public string? ImageFilename { get; private set; }
-    public decimal ProductLengthCm { get; private set; }
-    public decimal ProductWidthCm { get; private set; }
-    public decimal ProductHeightCm { get; private set; }
+    public float ProductLengthCm { get; private set; }
+    public float ProductWidthCm { get; private set; }
+    public float ProductHeightCm { get; private set; }
     public uint AvailableStock { get; private set; }
     public ProductCategory MainDepartment { get; private set; }
     public ProductCategory Subcategory { get; private set; }
@@ -34,7 +34,7 @@ public class CatalogItem
     public CatalogItem(
         string productName, string productDescription,
         decimal price, uint availableStock, ProductCategory subcategory,
-        decimal productLengthCm, decimal productWidthCm, decimal productHeightCm,
+        float productLengthCm, float productWidthCm, float productHeightCm,
         string sellerId, FulfillmentMethod fulfillmentMethod)
     {
         if (price <= 0m)
@@ -46,9 +46,9 @@ public class CatalogItem
         if (subcategory is null)
             throw new ArgumentNullException(nameof(subcategory));
 
-        var invalidArgName = productLengthCm <= 0m ? nameof(productLengthCm)
-            : productWidthCm <= 0m ? nameof(productWidthCm)
-            : productHeightCm <= 0m ? nameof(productHeightCm) : null;
+        var invalidArgName = productLengthCm <= 0f ? nameof(productLengthCm)
+            : productWidthCm <= 0f ? nameof(productWidthCm)
+            : productHeightCm <= 0f ? nameof(productHeightCm) : null;
         if (invalidArgName is not null)
         {
             throw new ArgumentOutOfRangeException(invalidArgName);
@@ -77,7 +77,7 @@ public class CatalogItem
         string productName, string productDescription,
         decimal price, string imageUri, uint availableStock,
         int mainDepartmentId, int subcategoryId,
-        decimal productLengthCm, decimal productWidthCm, decimal productHeightCm,
+        float productLengthCm, float productWidthCm, float productHeightCm,
         string sellerId, FulfillmentMethod fulfillmentMethod, bool hasOrdersInProgress)
     {
         ProductName = productName;
@@ -117,11 +117,11 @@ public class CatalogItem
         ImageFilename = imageFilename ?? ImageFilename;
     }
 
-    public void ChangeProductDimensions(decimal length, decimal width, decimal height)
+    public void ChangeProductDimensions(float length, float width, float height)
     {
-        var invalidArgName = length <= 0m ? nameof(length)
-            : width <= 0m ? nameof(width)
-            : height <= 0m ? nameof(height) : null;
+        var invalidArgName = length <= 0f ? nameof(length)
+            : width <= 0f ? nameof(width)
+            : height <= 0f ? nameof(height) : null;
         if (invalidArgName is not null)
         {
             throw new ArgumentOutOfRangeException(invalidArgName);
